ci: Enable provenance for npm package
This change adds the --provenance flag to npm publish.
We now benefit from attestation, that the published package on npm
was actually build on a GitHub Runner and not on somebodys local machine.
This is a great improvement for the trust relationship with users downloading the package and to increase supply-chain security for ziggy.

To do so, we also specify the permissions and allow the action to mint an unique ID token for the attestation, as described in the docs [0].

Additionally we also constrain the general permissions to the least possible.

Also fixed two yamllint indentation issues.

[0]: https://docs.npmjs.com/generating-provenance-statements#example-github-actions-workflow